Implemented parameter scoping
Priority is:

    session < export < carry < local (highest priority)

* session:
session parameters are set to a session, and used by all tasks in the
session.

* local:
local parameters are set to a task, and used only by the task.

* export:
export parameters are set to a task, and used by the task and its
subtasks.

* carry:
carry parameters are created by a task, and used by all following
subtasks and dependent tasks.