Transactional database secret store
INSERT into a table with unique index may fail with unique constraint
violation error even within a transaction that already has ensured that
the row doesn't exist using DELETE or SELECT. INSERT may run
concurrently because DELETE or SELECT doesn't lock the unexistent row.

A solution is to run DELETE and INSERT when a project is locked.