disable/enable schedule
Make it possible to mark a schedule as disabled, effectively suspending
or "pausing" a scheduled workflow until it is enabled again.

Scheduled sessions that have already been started are not affected.

When a schedule is re-enabled, it picks up where it stopped.
Sessions will not be skipped as a consequence of a disabling and later
enabling a schedule.