for_each: index subtask names
Instead of including the entire key and value in the names of generated
subtasks, use key and value indices. The key and value is still included
in the subtask name, but encoded and truncated so as to make it less
likely that the task name limit is hit.

This makes it possible to `for_each>` over arrays of larger records.

Another notable consequence of this change is that duplicate values in a
`for_each>` parameter array now generates separate subtasks. Previously
these duplicate values would generate subtasks with identical names.
Thus the below would previously only generate a single subtask.

```
for_each>: {a: [1, 1]}
``

The above will now generate two subtasks, as the user likely expects.