task timeouts
Track task start time and cancel the attempt if the time elapsed since
the first task start exceeds the task TTL.

Note that the TTL is not affected by retries. It is the first task
execution that is recorded as the task start time. A task cannot
circumvent the TTL by retrying.