Using _ instead of - for secret keys to be consistent with Config
Many secrets can be also set as parameters because the difference of
confidential and non-confidential is not always obvious for users. Some
may think AWS access_key_id is confidential but some may not.

Therefore, users and plugin developers may set the same configuration to
secrets or parameter differently, and their naming should be consistent
as much as possible to avoid misconfiguration, to make documents easier
to understand, and to make plugins easier to write.