Fix selfupdate on Windows
On Windows, a process can't change or delete a file if it is executing.
`selfupdate` command was failing due to this limitation. This change
uses a background process on Windows so that it repeats moving the file
until the execution finishes. The background process deletes itself at
the end using an exceptional behavior of Windows where a bat script can
delete itself.