Fix Guava version
`guava-retrying:2.0.0` depends on `guava:[10.+,)`. `:digdag-client`
depends on `guava-retrying:2.0.0` but doesn't depend on
guava directly.

In this case, apparently, gradle and maven solve the dependency
differently. Maven uses the latest guava across all dependencies to
guava (19.0 because dependency graph includes dependencies to guava:15,
16, 18 and 19). On the other hand, Gradle uses the latest available
version of guava (21.0 at this moment).

This change fixes this issue by changing dependency of digdag-client so
that it depends on guava:19 directly.