/api/attempts?project=&workflow= should lookup workflow by name
A workflow in a project is identified by projectId and name because
workflow ID changes every revision. Therefore, listing up attempts by
workflow name should not use workflow ID to query on the database.
Otherwise, before this fix, the REST API was listing up attempts that
were exected by the latest revision only.

Scheduler's `skip_on_overtime` option also had the same problem. It
should list up workflows by name so that workflows with the same name
shouldn't execute in parallel. It's unexpected if a workflow doesn't get
skipped when a new revision is pushed even though the workflow name or
definition is not changed.