Mitigate exit code 137 (OOM) on CI
Travis CI is occasionally causing this error:

    The command "travis_wait 40 ci/run_test.sh" exited with 137

Exit code 137 means 128 + 9 where 9 means SIGKILL. Likely this means
that the outer environment or kernel killed the process.
A possible cause is shortage of memory. Most of ./gradlew commands are
running with --no-daemon but some are not. This means that two gradlew
processes (daemon and non-daemon) run. Adding --no-daemon saves memory.