TransactionManager.getOrReuse for submodules that may or may not nested
Some modules such as NotificationSender may or may not in transaction
depending on its context. Accessing database from those modules have an
issue where it causes deadlock due to lack of connections in connection
pool, or causes "Nested transaction is not allowed" error.

TransactionManager.getOrReuse lets those modules share the same
transaction context with the external caller such as
WorkflowExecutionTimeoutEnforcer, while it can start a new transaction
if the external caller is not in transaction such as
NotifyOperatorFactory.

In fact, WorkflowExecutionTimeoutEnforcer should not call
NotificationSender synchronously. Instead, it should create a
notification task and send notification asynchronously. But here it's
postponed.