Fix CORS issue with direct log file download
PR #659 introduced a problem where browser rejects a request to get a
log file from a S3's pre-signed URL when S3 sends
"Access-Control-Allow-Origin: '*'" header.  This happens only when the
server generates a direct-download URL for log files.

Credentials should not be included in the request to accept
"Access-Control-Allow-Origin: '*'" header when the server generates
pre-signed URLs for log files.