Add cancelRequested flag to tasks in REST API
When a workflow is canceled, cancelRequested flag of attempt becomes
true in REST API. However, because tasks don't have cancelRequested
flag, it's relatively hard to reflect canceling state to the tasks on
console.

This change adds cancelRequested to tasks for console. Default value for
CLI clients is false for backward compatibility.