Add log-server.<type>.direct_download for workarounding CORS issues
In some environments, setting up CORS is not easy, or endpoint is not
reachable only from the digdag server. In this case, browsers fail to
download task log files using direct download URL.

This new option lets users to configure direct download as a part of
server configuration. Default is false.