Stop some cli commands call migrate to h2 memory db.
archive, push, start cli commands use DigdagEmbed.
In this case database migration is called and applied to empty h2 memory database.

This is no impcat to Digdag server becasuse migrate command requires the access to digdag database but no info is provied to these commands.
But after introducing migrate command, migrate info is displayed in console and it may be confused to users.
In addition we should not run useless migrate execution.
So in this PR, "database.migrate : false" set to DigdagEmbed and disable migration.