Remove duplicates when building resumed tasks map
When resuming a failed workflow, many tasks with the same full name
may be added, if many retries of the same task failed. This will throw
an uncaught exception when many tasks are added with the same
key (full name), to the map of tasks to resume.

This fixes the problem by discarding any duplicate tasks at the map
building stage. The assumption is that any tasks with the same name
are retries of the same task, so we only want one.