Support nested parens in JavaScript if engine is Graal
Config eval engine had a problem where JavaScript code can't use useful
functional methods such as Array.map because code couldn't include `}`
character. With this change, `}` is assumed as a part of JavaScript code
if number of `{` and `}` matches.

This change is enabled only for Graal JavaScript engine so that it
doesn't make unexpected backward incompatibility.