Implement #1266 in GraalJsEngine.
GraalEvaluator receive Context in constructor.
But createing Context and load js libraries are heavy process.
So it is better to delay it until evalute() called from the point of view of performances.
This commit change constructor argument to Supplier<Context> and
Context generate in evaluate().
It improve performance if the target code does not have ${..}.