Add option to copy symbolic links pointing outside of project directory (#1199)
* Add option to copy symbolic links pointing outside of project directory

`digdag push` fails if the project includes a symbolic link that points
outside of the project directory. This is hard to avoid because the
archive package must be self-contained but such files pointed by the
symlinks won't be included in the archive.

This new option provides a solution to it. If the option is given, the
command copies files or directories instead of throwing an exception.

* Better naming for --copy-outgoing-symlinks

* Fixed symlink handling at Archiver

When archiver creates a symbolic link pointing to a directory, files in
it shouldn't be copied. Such files should be copied through the
pointed directory, not though the symbolic link.

* fixed test cases