Not pass `engine_version` parameter if engine is presto
We manage dig file which includes a lot of `td` operators
and some jobs should be run by presto others should be hive.
Query engine of some jobs are determined dynamically.
In such a case we need to check engine and determine engine_version
if `engine_version` is passed for both presto and hive.

But we can not control presto engine_version via API job creation endpoint.
Then it's useless effort to determine engine_version.
We will need to specify `engine_version` of only hive by this change.